refactor: Move test file to correct folder
