+++ HerPortal/ExternalServices/CsvFiles/CsvFileGetter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using HerPortal.BusinessLogic.ExternalServices.CsvFiles;
 using HerPortal.BusinessLogic.Models;
 using HerPortal.BusinessLogic.ExternalServices.S3FileReader;
 using HerPortal.DataStores;
