+++ HerPortal/Startup.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using GovUkDesignSystem.ModelBinders;
 using Hangfire;
 using Hangfire.PostgreSql;
+using HerPortal.BusinessLogic.ExternalServices.CsvFiles;
 using HerPortal.BusinessLogic.ExternalServices.S3FileReader;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.DataProtection;
